FAAS/requirements.txt

Django==2.2
django-allauth==0.40.0
django-cors-headers==3.1.0
django-cors-middleware==1.5.0
django-oauth-toolkit==1.3.2
django-rest-auth==0.9.5
djangorestframework==3.10.3
six==1.14.0
sqlparse==0.3.0
gunicorn==20.0.4
django-heroku==0.3.1
whitenoise==5.1.0
requests==2.22.0
psycopg2==2.8.5
pytz==2018.9
urllib3==1.25.3
idna==2.8
dj-database-url==0.5.0
chardet==3.0.4